module-ize code
I guess back in the day it still wasn't common to serve a bundle, and so
you loaded stuff in a specific order and then attached the bits to some
app namespace (to keep out of global scope..).

A lot has changed.

I also removed skrollr, because it's not being supported anymore and who
knows what the module situation there is.